Update types and conversions

+++ Sand-Painter/Brush.cpp
@@ -9,7 +9,7 @@ Brush::Brush(int radius, int cell_size, sf::Vector2i centre)
 	m_radius_scaled(radius / cell_size), 
 	m_centre(centre), 
 	m_centre_scaled(centre / cell_size), 
-	m_cell_size(cell_size), m_circle(radius) 
+	m_cell_size(cell_size), m_circle(static_cast<float>(radius))
 {
 	m_circle.setFillColor(sf::Color::Transparent);
 	m_circle.setOutlineThickness(1.f);
